initialise tables

+++ src/main/java/gr/aueb/cf/eshopfinalproject/repository/ProductRepository.java
@@ -3,5 +3,5 @@ package gr.aueb.cf.eshopfinalproject.repository;
 import gr.aueb.cf.eshopfinalproject.model.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Products, Integer> {
+public interface ProductRepository extends JpaRepository<Products, Long> {
 }
